Refactor HyperbolicTargets: split into focused files, new namespace
Split HyperbolicTargets.cs (509 lines) into 5 focused files:
- HyperbolicTargets.cs: constants, helpers, ApplyPatches
- TransferPlannerPatches.cs: target list, Hohmann, ToF bounds, alignment
- EncounterPatches.cs: TryFindIntercept + closest approach sweep
- OrbitRenderingPatches.cs: asymptote clamping for orbit lines
- DiagnosticPatches.cs: debug transfer logging

Namespace: Features -> Features.HyperbolicTargets
